+++ script/install_dependencies.sh
@@ -9,5 +9,4 @@ bundle install --deployment
 RAILS_ENV=qa rake db:migrate
 yarn install
 RAILS_ENV=qa WEBPACKER_PRECOMPILE=false rake assets:precompile
-RAILS_ENV=qa rake pillar_theme:webpacker:compile
-# RAILS_ENV=qa WEBPACKER_PRECOMPILE=false rake assets:clean
+RAILS_ENV=qa bundle exec rake --trace pillar_theme:webpacker:compile