+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 ### PENDING RELEASE (TBD)
-
+- Update doc signing calls to EJBCA to remove email addresses and use different EJBCA profiles
 - Redirect to initially requested URL, after login and DUO authentication
 
 ### v1.5.0 (August 4, 2020)