+++ CHANGELOG.md
@@ -3,9 +3,12 @@ All notable changes to this project will be documented in this file.
 
 ### PENDING RELEASE (TBD)
 - Add Prometheus endpoints for metrics
+- Rename Weak Keys to Reject Keys
 - Update Code Deploy AWS configuration
 - Update doc signing calls to EJBCA to remove email addresses and use different EJBCA profiles
 - Redirect to initially requested URL, after login and DUO authentication
+- Fix button alignment on the registrant edit screen
+- fix csr verify signature
 
 ### v1.5.0 (August 4, 2020)